* 'master' of github.com:scalableminds/webknossos: (337 commits)
  Fix docs for the annotation download file format (#6546)
  Added total runtime information to VX reports (#6543)
  fix VX report for completed + skipped tasks (#6540)
  Avoid allocating spire uint objects during apply agglomerate (#6532)
  Explore remote N5 datasets (#6520)
  Fix MeshChunk byteOffset (Long, not Int) (#6536)
  update browserslist (#6505)
  Support new Mesh File (v3) (#6491)
  makes workflow_yamlContent optional (#6518)
  Always return 404 for Failures in Zarr Streaming (#6515)
  Poll wk version to notify during upgrade (#6451)
  add script which extracts newest changelog and creates GH release for it (#6504)
  release 22.10.0 (#6500)
  voxel³ -> voxel (#6501)
  Allow task type summary to identify task type when creating tasks in bulk (#6486)
  Fix sql evolution 090 (defer not null constraint) (#6498)
  SQL schema cleanup (#6492)
  Fix validation of layer selection when trying to start globalization of floodfills (#6497)
  Add "shift + w" shortcut to cycle backwards through tools (#6493)
  Fix filtering for public datasets in dataset table (#6496)
  ...

+++ tools/proxy/proxy.js
@@ -13,7 +13,7 @@ const app = express();
 
 const ROOT = path.resolve(path.join(__dirname, "..", ".."));
 const PORT = parseInt(process.env.PORT || 9000, 10);
-const HOST = `http://localhost:${PORT}`;
+const HOST = `http://127.0.0.1:${PORT}`;
 const loggingPrefix = "Proxy:";
 
 function makeEnv(port, host) {
@@ -69,19 +69,21 @@ function spawnIfNotSpecified(keyword, command, args, options) {
 
 function shutdown() {
   console.log("", loggingPrefix, "Shutting down, terminating child processes...");
-  for (const proc of Object.values(processes).filter(x => x)) {
+  for (const proc of Object.values(processes).filter((x) => x)) {
     if (proc.connected) {
       proc.kill("SIGTERM");
     }
   }
-  exec("kill $(lsof -t -i:5005)"); // Also kill Java debug subproces, as it’s sometimes not terminated by sbt properly.
+  if (processes.backend != null) {
+    exec("kill $(lsof -t -i:5005)"); // Also kill Java debug subprocess, as it’s sometimes not terminated by sbt properly.
+  }
   process.exit(0);
 }
 
-for (const [key, proc] of Object.entries(processes).filter(x => x[1] !== null)) {
+for (const [key, proc] of Object.entries(processes).filter((x) => x[1] !== null)) {
   proc.stdout.pipe(prefixLines(`${key}: `)).pipe(process.stdout);
   proc.stderr.pipe(prefixLines(`${key}: `)).pipe(process.stderr);
-  proc.on("error", err => console.error(err, err.stack));
+  proc.on("error", (err) => console.error(err, err.stack));
   proc.on("exit", shutdown);
 }
 process.on("SIGTERM", shutdown);
@@ -96,11 +98,11 @@ proxy.on("error", (err, req, res) => {
 });
 
 function toBackend(req, res) {
-  proxy.web(req, res, { target: `http://localhost:${PORT + 1}` });
+  proxy.web(req, res, { target: `http://127.0.0.1:${PORT + 1}` });
 }
 
 function toWebpackDev(req, res) {
-  proxy.web(req, res, { target: `http://localhost:${PORT + 2}` });
+  proxy.web(req, res, { target: `http://127.0.0.1:${PORT + 2}` });
 }
 
 app.all("/assets/bundle/*", toWebpackDev);
